Extract GeneralSettingItem
